+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Current Master
 
+## 4.11.1 Wed Mar 20 2019
+
+- [#266](https://github.com/poanetwork/nifty-wallet/pull/266): (Fix) Support of the latest trezor-connect version 7
+- [#263](https://github.com/poanetwork/nifty-wallet/pull/263): (Refactoring) Refine dropdown components
+
 ## 4.11.0 Fri Feb 08 2019
 
 ### Features